api/requirements.txt
pylint
pytest
Flask
python-dotenv
psycopg2